🔧 Move jupytext/kernelspec frontmatter code
