+++ packages/myst-frontmatter/src/jupytext/types.ts
@@ -1,0 +1,11 @@
+export type TextRepresentation = {
+  extension?: string;
+  format_name?: string;
+  format_version?: string;
+  jupytext_version?: string;
+};
+
+export type Jupytext = {
+  formats?: string;
+  text_representation?: TextRepresentation;
+};
