+++ packages/myst-frontmatter/src/kernelspec/types.ts
@@ -1,0 +1,7 @@
+export type KernelSpec = {
+  name?: string;
+  language?: string;
+  display_name?: string;
+  argv?: string[];
+  env?: Record<string, any>;
+};
