+++ packages/myst-frontmatter/src/page/validators.ts
@@ -3,7 +3,6 @@ import {
   defined,
   incrementOptions,
   validateList,
-  validateObject,
   validateObjectKeys,
   validateString,
 } from 'simple-validators';
@@ -11,8 +10,10 @@ import {
   PROJECT_AND_PAGE_FRONTMATTER_KEYS,
   validateProjectAndPageFrontmatterKeys,
 } from '../project/validators.js';
-import type { Jupytext, KernelSpec, PageFrontmatter, TextRepresentation } from './types.js';
+import type { PageFrontmatter } from './types.js';
 import { FRONTMATTER_ALIASES } from '../index.js';
+import { validateKernelSpec } from '../kernelspec/validators.js';
+import { validateJupytext } from '../jupytext/validators.js';
 
 export const PAGE_FRONTMATTER_KEYS = [
   ...PROJECT_AND_PAGE_FRONTMATTER_KEYS,
@@ -40,91 +41,6 @@ export const USE_PROJECT_FALLBACK = [
   'funding',
   'affiliations',
 ];
-
-const KERNELSPEC_KEYS = ['name', 'language', 'display_name', 'argv', 'env'];
-const TEXT_REPRESENTATION_KEYS = ['extension', 'format_name', 'format_version', 'jupytext_version'];
-const JUPYTEXT_KEYS = ['formats', 'text_representation'];
-
-/**
- * Validate KernelSpec object
- */
-export function validateKernelSpec(input: any, opts: ValidationOptions) {
-  const value = validateObjectKeys(input, { optional: KERNELSPEC_KEYS }, opts);
-  if (value === undefined) return undefined;
-  const output: KernelSpec = {};
-  if (defined(value.name)) {
-    output.name = validateString(value.name, incrementOptions('name', opts));
-  }
-  if (defined(value.language)) {
-    output.language = validateString(value.language, incrementOptions('language', opts));
-  }
-  if (defined(value.display_name)) {
-    output.display_name = validateString(
-      value.display_name,
-      incrementOptions('display_name', opts),
-    );
-  }
-  if (defined(value.env)) {
-    output.env = validateObject(value.env, incrementOptions('env', opts));
-  }
-  if (defined(value.argv)) {
-    output.argv = validateList(value.argv, incrementOptions('argv', opts), (arg, index) => {
-      return validateString(arg, incrementOptions(`argv.${index}`, opts));
-    });
-  }
-  return output;
-}
-
-function validateTextRepresentation(input: any, opts: ValidationOptions) {
-  const value = validateObjectKeys(input, { optional: TEXT_REPRESENTATION_KEYS }, opts);
-  if (value === undefined) return undefined;
-  const output: TextRepresentation = {};
-  if (defined(value.extension)) {
-    output.extension = validateString(value.extension, incrementOptions('extension', opts));
-  }
-  if (defined(value.format_name)) {
-    output.format_name = validateString(value.format_name, incrementOptions('format_name', opts));
-  }
-  if (defined(value.format_version)) {
-    // The format version occasionally comes as a number in YAML, treat it as a string
-    const format_version =
-      typeof value.format_version === 'number'
-        ? String(value.format_version)
-        : value.format_version;
-    output.format_version = validateString(
-      format_version,
-      incrementOptions('format_version', opts),
-    );
-  }
-  if (defined(value.jupytext_version)) {
-    output.jupytext_version = validateString(
-      value.jupytext_version,
-      incrementOptions('jupytext_version', opts),
-    );
-  }
-  return output;
-}
-
-/**
- * Validate Jupytext object
- *
- * https://jupyterbook.org/en/stable/file-types/myst-notebooks.html
- */
-export function validateJupytext(input: any, opts: ValidationOptions) {
-  const value = validateObjectKeys(input, { optional: JUPYTEXT_KEYS }, opts);
-  if (value === undefined) return undefined;
-  const output: Jupytext = {};
-  if (defined(value.formats)) {
-    output.formats = validateString(value.formats, incrementOptions('formats', opts));
-  }
-  if (defined(value.text_representation)) {
-    output.text_representation = validateTextRepresentation(
-      value.text_representation,
-      incrementOptions('text_representation', opts),
-    );
-  }
-  return output;
-}
 
 export function validatePageFrontmatterKeys(value: Record<string, any>, opts: ValidationOptions) {
   const output: PageFrontmatter = validateProjectAndPageFrontmatterKeys(value, opts);
